* Further misc code cleanups after C++17 switch
* (Re?)Add windows msys2 llvm ci coverage
* Re-fix black level fixup for Panasonic cameras, only for version 5 or later.
* Ricoh GR II firmware 3.0 alias
* Add Sony DSC-HX95 alias
* Add Leica M10 normalization
* Leica DIGILUX 3 support
* More Leica C support

* upstream/develop: (93 commits)
  More Leica C support
  cameras.xml: Fix camera make in <ID> tag, it's Leica, not Panasonic
  Rw2Decoder: there may not be a version 'PANASONIC_RAWFORMAT' tag
  rawspeed: Leica DIGILUX 3 support
  Add Leica M10 normalization
  Rw2Decoder::decodeMetaDataInternal(): getBlack(): early return
  [CI] Update to LLVM13, add LLVM12 into legacy test set
  [CMake] When building for APPLE, and not targeting 11.3+, don't even look for `aligned_alloc()`
  [CI] Add windows msys2 llvm jobs
  Update linker flags for Windows
  Handle LLVM13 build error
  Add Sony DSC-HX95 alias
  Ricoh GR II firmware 3.0 alias
  appease the static compliation analysis...
  Update Rw2Decoder.cpp
  fix capture list.
  remove auto.
  Pivot black levels on raw version. Newer then version 5 should have correct black levels.
  Revert "CameraMetaData: use structured bindings"
  Partially revert "Mark maybe-unused function arguments as such"
  ...

+++ src/librawspeed/io/Endianness.h
@@ -70,18 +70,12 @@ inline Endianness getHostEndianness() {
 inline int16_t getByteSwapped(int16_t v) {
   return static_cast<int16_t>(BSWAP16(static_cast<uint16_t>(v)));
 }
-inline uint16_t getByteSwapped(uint16_t v) {
-  return static_cast<uint16_t>(BSWAP16(v));
-}
+inline uint16_t getByteSwapped(uint16_t v) { return BSWAP16(v); }
 inline int32_t getByteSwapped(int32_t v) {
   return static_cast<int32_t>(BSWAP32(static_cast<uint32_t>(v)));
 }
-inline uint32_t getByteSwapped(uint32_t v) {
-  return static_cast<uint32_t>(BSWAP32(v));
-}
-inline uint64_t getByteSwapped(uint64_t v) {
-  return BSWAP64(static_cast<uint64_t>(v));
-}
+inline uint32_t getByteSwapped(uint32_t v) { return BSWAP32(v); }
+inline uint64_t getByteSwapped(uint64_t v) { return BSWAP64(v); }
 
 // the float/double versions use two memcpy which guarantee strict aliasing
 // and are compiled into the same assembly as the popular union trick.
